Create -t|--template option
The template option will create a new template file in the current
directory and exit. This may be useful for moving an older script into
the newer layout.

Copy the usage, main and local functions from the old script to files in
the project directory. Then copy them into the new template.